Enhance API key management and refine UI elements across the application
Update API key schema and backend services to support new fields like environment tagging and cost limits. Modify frontend components for the API Keys, Policy Designer, Guardrails, Router Config, and Settings pages to incorporate these enhancements and improve user experience by making account limits editable and refining UI design. Adjust API endpoints and documentation accordingly.

+++ backend/app/db/models/api_key.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, JSON
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, JSON, Float
 from sqlalchemy.orm import relationship
 from datetime import datetime
 
@@ -26,8 +26,8 @@ class APIKey(Base):
     allowed_models_override = Column(JSON, nullable=True)
     allowed_providers_override = Column(JSON, nullable=True)
     
-    cost_limit_daily = Column(Integer, nullable=True)
-    cost_limit_monthly = Column(Integer, nullable=True)
+    cost_limit_daily = Column(Float, nullable=True)
+    cost_limit_monthly = Column(Float, nullable=True)
     
     metadata_ = Column("metadata", JSON, default=dict)
